+++ lolProject/lolLib/DTO/Game.cs
@@ -9,9 +9,9 @@
         public String platformId { get; set; }
         public Int64 gameCreation { get; set; }
         public Int32 gameDuration { get; set; }
-        public Int32 queueId { get; set; }
-        public Int32 mapId { get; set; }
-        public Int32 seasonId { get; set; }
+        public Int16 queueId { get; set; }
+        public Int16 mapId { get; set; }
+        public Int16 seasonId { get; set; }
         public String gameVersion { get; set; }
         public String gameMode { get; set; }
         public String gameType { get; set; }
