+++ lolProject/lolLib/DTO/Mastery.cs
@@ -5,6 +5,6 @@
     public class Mastery : IDTO
     {
         public Int32 masteryId { get; set; }
-        public Int32 rank { get; set; }
+        public SByte rank { get; set; }
     }
 }
