+++ lolProject/lolLib/DTO/Participant.cs
@@ -6,8 +6,8 @@
     public class Participant : IDTO
     {
         public Int32 participantId { get; set; }
-        public Int32 teamId { get; set; }
-        public Int32 championId { get; set; }
+        public SByte teamId { get; set; }
+        public Int16 championId { get; set; }
         public Int32 spell1Id { get; set; }
         public Int32 spell2Id { get; set; }
         public List<Mastery> masteries { get; set; }
