+++ lolProject/lolLib/DTO/ParticipantFrame.cs
@@ -8,10 +8,10 @@
         public Position position { get; set; }
         public Int32 currentGold { get; set; }
         public Int32 totalGold { get; set; }
-        public Int32 level { get; set; }
+        public SByte level { get; set; }
         public Int32 xp { get; set; }
-        public Int32 minionsKilled { get; set; }
-        public Int32 jungleMinionsKilled { get; set; }
+        public SByte minionsKilled { get; set; }
+        public SByte jungleMinionsKilled { get; set; }
         public Int32 dominionScore { get; set; }
         public Int32 teamScore { get; set; }
     }
