+++ lolProject/lolLib/DTO/Rune.cs
@@ -5,6 +5,6 @@
     public class Rune : IDTO
     {
         public Int32 runeId { get; set; }
-        public Int32 rank { get; set; }
+        public Int16 rank { get; set; }
     }
 }
